\name{gena-package}
\alias{gena-package}
\alias{gena}
\docType{package}
\title{
\packageTitle{gena}
}
\description{
\packageDescription{gena}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gena}
\packageIndices{gena}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{gena}

Maintainer: \packageMaintainer{gena}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
