% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gena.R
\name{gena}
\alias{gena}
\title{Genetic Algorithm}
\usage{
gena(
  fn,
  gr = NULL,
  lower,
  upper,
  pop.n = 100,
  pop.initial = NULL,
  mating.method = "rank",
  mating.par = NULL,
  mating.self = FALSE,
  crossover.method = "local",
  crossover.par = NULL,
  crossover.prob = 0.8,
  mutation.method = "percent",
  mutation.par = NULL,
  mutation.prob = 0.2,
  mutation.genes.prob = 1/length(lower),
  elite.n = min(10, 2 * round(pop.n/20)),
  elite.duplicates = FALSE,
  hybrid.method = "rank",
  hybrid.par = 2,
  hybrid.prob = 0.1,
  hybrid.opt.par = NULL,
  hybrid.n = 1,
  maxiter = 100,
  constr = NULL,
  info = TRUE,
  ...
)
}
\arguments{
\item{fn}{function to be maximized}

\item{gr}{gradient of the \code{fn}}

\item{lower}{lower bound of the search space}

\item{upper}{upper bound of the search space}

\item{pop.n}{integer representing the size of the population}

\item{pop.initial}{numeric matrix which rows are chromosomes to be
included into the initial population}

\item{mating.method}{the algorithm to be applied for a mating i.e. selection
of parents}

\item{mating.par}{parameters of the mating algorithm (see details)}

\item{mating.self}{logical; if \code{TRUE} then the chromosome may
mate with itself i.e. both parents may be the same chromosome}

\item{crossover.method}{an algorithm to be applied for crossover i.e.
creation of the children}

\item{crossover.par}{parameters of the crossover algorithm (see details)}

\item{crossover.prob}{probability of the crossover for each pair of parents}

\item{mutation.method}{algorithm to be applied for mutation i.e. random
change in some genes of the children}

\item{mutation.par}{parameters of the mutation algorithm (see details)}

\item{mutation.prob}{mutation probability for the chromosomes}

\item{mutation.genes.prob}{mutation probability for the genes}

\item{elite.n}{number of the elite children i.e. those which have the
highest function value and will be preserved for the next population}

\item{elite.duplicates}{logical; if \code{TRUE} then some elite children
may have the same genes}

\item{hybrid.method}{hybrids selection algorithm i.e. mechanism 
determining which chromosomes should be subject to local optimization}

\item{hybrid.par}{parameters of the hybridization algorithm (see details)}

\item{hybrid.prob}{probability of generating the hybrids each iteration}

\item{hybrid.opt.par}{parameters of the local optimization function
to be used for hybridization algorithm}

\item{hybrid.n}{number of hybrids that appear if hybridization
should take place during the iteration}

\item{maxiter}{maximum number of iterations of the algorithm}

\item{constr}{a function imposing a constraints on each of the chromosomes}

\item{info}{logical; if \code{TRUE} then some optimization related 
information will be printed each iteration}

\item{...}{additional parameters to be passed to 
\code{fn}, \code{gr} and \code{constr} functions}
}
\value{
This function returns the list containing the information on the
results of the genetic algorithm
}
\description{
This function allows to use genetic algorithm for
numeric global optimization purposes.
}
