% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gena.R
\name{plot.gena}
\alias{plot.gena}
\title{Plot best found fitnesses during genetic algorithm}
\usage{
\method{plot}{gena}(x, y = NULL, ...)
}
\arguments{
\item{x}{Object of class "gena"}

\item{y}{this parameter currently ignored}

\item{...}{further arguments (currently ignored)}
}
\value{
This function does not return anything.
}
\description{
Plot best found fitnesses during genetic algorithm
}
