% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pso.R
\name{plot.pso}
\alias{plot.pso}
\title{Plot best found fitnesses during genetic algorithm}
\usage{
\method{plot}{pso}(x, y = NULL, ...)
}
\arguments{
\item{x}{Object of class "pso"}

\item{y}{this parameter currently ignored}

\item{...}{further arguments (currently ignored)}
}
\value{
This function does not return anything.
}
\description{
Plot best found fitnesses during genetic algorithm
}
