% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gena.R
\name{print.gena}
\alias{print.gena}
\title{Print method for "gena" object}
\usage{
\method{print}{gena}(x, ...)
}
\arguments{
\item{x}{Object of class "gena"}

\item{...}{further arguments (currently ignored)}
}
\value{
This function does not return anything.
}
\description{
Print method for "gena" object
}
