% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pso.R
\name{print.pso}
\alias{print.pso}
\title{Print method for "pso" object}
\usage{
\method{print}{pso}(x, ...)
}
\arguments{
\item{x}{Object of class "pso"}

\item{...}{further arguments (currently ignored)}
}
\value{
This function does not return anything.
}
\description{
Print method for "pso" object
}
