% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gena.R
\name{print.summary.gena}
\alias{print.summary.gena}
\title{Summary for "gena" object}
\usage{
\method{print}{summary.gena}(x, ...)
}
\arguments{
\item{x}{Object of class "gena"}

\item{...}{further arguments (currently ignored)}
}
\value{
This function returns \code{x} input argument.
}
\description{
Summary for "gena" object
}
