% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pso.R
\name{print.summary.pso}
\alias{print.summary.pso}
\title{Summary for "pso" object}
\usage{
\method{print}{summary.pso}(x, ...)
}
\arguments{
\item{x}{Object of class "pso"}

\item{...}{further arguments (currently ignored)}
}
\value{
This function returns \code{x} input argument.
}
\description{
Summary for "pso" object
}
