% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gena.R
\name{summary.gena}
\alias{summary.gena}
\title{Summarizing gena Fits}
\usage{
\method{summary}{gena}(object, ...)
}
\arguments{
\item{object}{Object of class "gena"}

\item{...}{further arguments (currently ignored)}
}
\value{
This function returns the same list as \code{\link[gena]{gena}} 
function changing its class to "summary.gena".
}
\description{
Summarizing gena Fits
}
