% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pso.R
\name{summary.pso}
\alias{summary.pso}
\title{Summarizing pso Fits}
\usage{
\method{summary}{pso}(object, ...)
}
\arguments{
\item{object}{Object of class "pso"}

\item{...}{further arguments (currently ignored)}
}
\value{
This function returns the same list as \code{\link[gena]{pso}} 
function changing its class to "summary.pso".
}
\description{
Summarizing pso Fits
}
